\name{retweetStatus}
\alias{retweetStatus}

\title{
Retweet's a selected tweet pulled down from twitteR
}
\description{
Retweet tweets useful for bot retweeting
}
\usage{
retweetStatus(status, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{status}{
The complete status to be retweeted
}
}

\value{
\item{TRUE}{If the status is successfully retweeted}
\item{FALSE}{If the status fails to retweet}
}

\author{
Kevin Brosnan
}
